pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-mesh-router"
version = "0.1.0"
description = "Distributed multi-agent task orchestration router with SQLite persistence"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "aiosqlite>=0.19.0",
    "cloudevents>=1.12.0",
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "sdnotify>=0.3.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
src = ["src", "tests"]
exclude = ["kiss_mesh", ".claude-flow", ".swarm", ".claude"]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.mypy]
files = ["src"]
ignore_missing_imports = true

[tool.pyright]
include = ["src"]
exclude = ["kiss_mesh", "tests", ".claude-flow", ".swarm", ".claude"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
